+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Errors returned by this package provide also optional details map accessible
   through `detailer` interface.
+- `WithDetails` which wraps an error exposing access to (a potentially new layer of)
+  details about the error.
 
 ## [0.3.0] - 2022-01-03
 
@@ -31,7 +33,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - First public release.
 
-[Unreleased]: https://gitlab.com/tozd/go/errors/-/compare/v0.3.0...main
+[unreleased]: https://gitlab.com/tozd/go/errors/-/compare/v0.3.0...main
 [0.3.0]: https://gitlab.com/tozd/go/errors/-/compare/v0.2.0...v0.3.0
 [0.2.0]: https://gitlab.com/tozd/go/errors/-/compare/v0.1.0...v0.2.0
 [0.1.0]: https://gitlab.com/tozd/go/errors/-/tags/v0.1.0